// src/types/Page.ts
export type Blade =
  | Agenda
  | Author
  | Cards
  | Comparisontable
  | Contactus
  | Contiguouscontainer
  | Faq
  | Footnote
  | Form
  | Helloworld
  | Horizontaltabs
  | Index
  | Logogrid
  | Marquee
  | Media1
  | Notification
  | Nup
  | Nupfixed
  | Offer
  | Oneup
  | Open
  | Productfeature
  | Productportfolio
  | Promotion
  | Simpletable
  | Story
  | Summary
  | Tableofcontents
  | Text
  | Textlongform
  | Textmainbody
  | Textwithimage
  | Trailheadmfe
  | Videoblade;
export type Background = Custom | Standard | Swoosh | Container;
// ...
export type Blades = Blade[];

export interface Page {
  nav?: Nav;
  context_nav?: ContextNav;
  blades?: Blades;
  pageAttrs?: {
    read_time?: string;
    livechat?: {
      enabled?: boolean;
      headline?: string;
      name_capture_campaign_id?: string;
      use_default_prompt?: boolean;
      prompts?: string[];
      [k: string]: unknown;
    };
    agentforce?: {
      org_id?: string;
      deployment_dev_name?: string;
      messaging_url?: string;
      script_url?: string;
      [k: string]: unknown;
    };
    content_type?: string;
    seo_page_type?: string;
    seo_product_name?: string;
    alt_seo_product_name?: string;
    autogenerate_json_ld?: boolean;
    page_theme_variant?: string;
    page_background_type?: PageBackground;
    [k: string]: unknown;
  };
  taxonomy?: {
    pageTagsAuto?: Tag[];
    pageTagsManual?: Tag[];
    [k: string]: unknown;
  };
  templateId?: string;
  templateName?: string;
  theme?: string;
  globalNavEnabled?: boolean;
  footerEnabled?: boolean;
  footer?: Footer;
  [k: string]: unknown;
}
export interface Nav {
  showRegion?: boolean;
  showAccount?: boolean;
  showSearch?: boolean;
  showContact?: boolean;
  showCustomerOffer?: boolean;
  showProspectOffer?: boolean;
  customerOfferOverride?: Link;
  prospectOfferOverride?: Link;
  headerType?: string;
  [k: string]: unknown;
}
export interface Link {
  /**
   * The label of the button.
   */
  label?: string;
  typeOptions?: {
    [k: string]: unknown;
  };
  ariaLabel?: string;
  url?: string;
  target?: "_self" | "_parent" | "_top" | "_blank";
  isSpa?: boolean;
  [k: string]: unknown;
}
/**
 * Context navigation is part of the overall navigation system and works in conjuction with top navigation
 */
export interface ContextNav {
  newWindowAlt?: string;
  variation: "static" | "dynamic" | "minimized" | "standalone";
  /**
   * Extra spacing to vertically move the navigation items up (by specifying a negative value i.e. '-10px') or down (by specifying a positive value i.e. '10px').
   */
  navTopExtraSpacing?: string;
  /**
   * For the standalone variation only: Desired width of the logo used in the navigation i.e. '156px'.
   */
  standaloneLogoWidth?: string;
  /**
   * For the standalone variation only: If true, centers the navigation items, otherwise navigation items are left-aligned by default.
   */
  standaloneCentered?: boolean;
  isLogoEnabled: boolean;
  propertyLogo?: Image;
  propertyTitle: Link;
  menuGroup: ContextNavMenugroup;
  utilityNavigation?: ContextNavUtilityNav;
  orgSelector?: ContextNavOrgSelector;
  [k: string]: unknown;
}
export interface Image {
  kind?: "image";
  title?: string | null;
  src: string;
  width: number | string;
  height: number | string;
  alt?: string;
  hint?: string;
  is_decorative?: boolean;
  typeOptions?: {
    [k: string]: unknown;
  };
  sizes?: Size[];
  image_background_circle_color?: string;
  image_background_gradient_color?: string;
  image_alignment?: string;
  file_url?: string;
}
export interface Size {
  src?: string;
  width?: number;
  height?: number;
  [k: string]: unknown;
}
/**
 * Group of menu items
 */
export interface ContextNavMenugroup {
  ID?: number;
  eyebrowHeading?: string;
  /**
   * @minItems 1
   */
  menuItems: ContextNavMenuitem[];
  ctaLink?: Link;
  hasSeparator?: boolean;
  [k: string]: unknown;
}
/**
 * Context menu utility navigation items
 */
export interface ContextNavUtilityNav {
  ctamain?: {
    cta?: Cta;
    inMobileHeader?: boolean;
    isSpa?: boolean;
    [k: string]: unknown;
  };
  ctasecondary?: {
    cta?: Cta;
    inMobileHeader?: boolean;
    isSpa?: boolean;
    [k: string]: unknown;
  };
  [k: string]: unknown;
}
// ...
export interface ContextNavOrgSelector {
  selectorLabel?: string;
  title?: string;
  /**
   * @minItems 1
   */
  orgList?: ContextNavOrg[];
  footer?: Link;
  [k: string]: unknown;
}
/**
 * Context nav organization list
 */
export interface ContextNavOrg {
  ID?: number;
  link?: Link;
  email?: string;
  action?: string;
  icon?: {
    src?: string;
    alt?: string;
    [k: string]: unknown;
  };
  helpbubble?: string;
  description?: string;
  logo?: {
    src?: string;
    alt?: string;
    [k: string]: unknown;
  };
  [k: string]: unknown;
}
export interface Agenda {
  attributes?: {
    audience_segment?: string[];
    theme_variant?: string;
    minimize_blade_spacing_bottom?: boolean;
    minimize_blade_spacing?: boolean;
    experiment_id?: string;
    experiment_type?: string;
    eyebrow?: Eyebrow;
    headline: string;
    headline_font_color?: string;
    subheadline?: string;
    description?: string;
    anchor_id?: string;
    link_1?: Cta;
    link_2?: Cta;
    blade_background_type?: Background;
    esi_blade_flag?: string;
    esi_blade_variant?: string;
    esi_blade_rule?: string;
  };
  name: "sf/agenda";
  id: string;
  tab_items?: CardAgendaCategory[];
  [k: string]: unknown;
}
export interface Eyebrow {
  label?: string;
  image?: Image;
  icon?: Icon;
  info_bubble?: string;
  hint?: string;
  character_count?: string;
  [k: string]: unknown;
}
export interface Icon {
  kind?: "icon";
  label?: string;
  src: string;
  decorative?: boolean;
  ariaLabel?: string;
  hint?: string;
}
export interface Custom {
  kind: "custom_background";
  blade_background_color?: string;
  blade_custom_image_desktop: Image;
  blade_custom_image_mobile: Image;
  blade_background_gradient_direction?: string;
}
export interface Standard {
  kind: "standard_background";
  blade_background_color?: string;
  blade_background_image?: Image;
  blade_background_shape?: string;
  blade_background_gradient_direction?: string;
}
export interface Swoosh {
  kind: "swoosh_background";
  blade_background_color: string;
}
export interface Container {
  kind: "container_background";
  blade_background_color?: string;
  blade_background_image?: Image;
  blade_background_image_position?: string;
  blade_background_shape?: string;
  blade_background_gradient_direction?: string;
}
// ...
export interface Author {
  attributes?: {
    audience_segment?: string[];
    theme_variant?: string;
    minimize_blade_spacing?: boolean;
    minimize_blade_spacing_bottom?: boolean;
    experiment_id?: string;
    experiment_type?: string;
    author_image?: Image;
    full_name: string;
    headline?: string;
    description: string;
    corner_image?: Image;
    link_1?: Cta;
    blade_background_type?: Background;
    esi_blade_flag?: string;
    esi_blade_variant?: string;
    esi_blade_rule?: string;
  };
  name: "sf/author";
  id: string;
  [k: string]: unknown;
}
// ...
export interface Tag {
  uuid: string;
  tagName: string;
  tagUuid: string;
}
// ...
export interface Comparisontable {
  attributes?: {
    audience_segment?: string[];
    theme_variant?: string;
    minimize_blade_spacing_bottom?: boolean;
    minimize_blade_spacing?: boolean;
    experiment_id?: string;
    experiment_type?: string;
    headline: string;
    headline_font_color?: string;
    anchor_id?: string;
    ucs_content_id: ExternalRef;
    link_1?: Link;
    link_2?: Link;
    blade_background_type?: Background;
    esi_blade_flag?: string;
    esi_blade_variant?: string;
    esi_blade_rule?: string;
  };
  name: "sf/comparisontable";
  id: string;
  media_items?: FeatureComparisonGroup[];
  [k: string]: unknown;
}
export interface FeatureComparisonGroup {
  id: string;
  name: string;
  /**
   * @minItems 1
   * @maxItems 20
   */
  features: {
    name: string;
    description?: string;
    editionStatus?: {
      editionId: string;
      editionStatus: string;
      overrideText?: string;
      [k: string]: unknown;
    }[];
    [k: string]: unknown;
  }[];
}
export interface Contactus {
  attributes?: {
    audience_segment?: string[];
    theme_variant?: string;
    minimize_blade_spacing_bottom?: boolean;
    minimize_blade_spacing?: boolean;
    experiment_id?: string;
    experiment_type?: string;
    headline: string;
    headline_font_color?: string;
    anchor_id?: string;
    blade_background_type?: Background;
    esi_blade_flag?: string;
    esi_blade_variant?: string;
    esi_blade_rule?: string;
  };
  name: "sf/contactus";
  id: string;
  media_items?: CardContactUs[];
  [k: string]: unknown;
}
export interface Contiguouscontainer {
  attributes?: {
    audience_segment?: string[];
    theme_variant?: string;
    anchor_id?: string;
    blade_background_type?: Background;
  };
  name: "sf/contiguouscontainer";
  id: string;
  media_items?: ChildrenContiguousContainer[];
  [k: string]: unknown;
}
export interface ChildrenContiguousContainer {
  id: string;
  label: string;
  kind: "children_contiguous-container";
  blades: (
    | Agenda
    | Author
    | Comparisontable
    | Contactus
    | Faq
    | Footnote
    | Index
    | Logogrid
    | Marquee
    | Media1
    | Notification
    | Nup
    | Nupfixed
    | Offer
    | Oneup
    | Productfeature
    | Productportfolio
    | Promotion
    | Simpletable
    | Story
    | Summary
    | Text
    | Textlongform
    | Textmainbody
    | Trailheadmfe
  )[];
}
export interface Faq {
  attributes?: {
    audience_segment?: string[];
    theme_variant?: string;
    minimize_blade_spacing_bottom?: boolean;
    minimize_blade_spacing?: boolean;
    experiment_id?: string;
    experiment_type?: string;
    headline: string;
    headline_font_color?: string;
    anchor_id?: string;
    blade_background_type?: Background;
    esi_blade_flag?: string;
    esi_blade_variant?: string;
    esi_blade_rule?: string;
  };
  name: "sf/faq";
  id: string;
  tab_items?: CardFAQCategory[];
  [k: string]: unknown;
}
export interface CardFAQCategory {
  id: string;
  kind: "card_faq-category";
  label: string;
  /**
   * @minItems 1
   * @maxItems 4
   */
  content: AccordionGroup[];
}
export interface AccordionGroup {
  id?: string;
  kind: "accordion_group";
  accordions: {
    headline: string;
    description: string;
    [k: string]: unknown;
  }[];
}
export interface Footnote {
  attributes?: {
    blade_layout?: string;
    audience_segment?: string[];
    theme_variant?: string;
    minimize_blade_spacing_bottom?: boolean;
    minimize_blade_spacing?: boolean;
    experiment_id?: string;
    experiment_type?: string;
    footnote_text: string;
    anchor_id?: string;
    column_grid?: string;
    esi_blade_flag?: string;
    esi_blade_variant?: string;
    esi_blade_rule?: string;
  };
  name: "sf/footnote";
  id: string;
  [k: string]: unknown;
}
export interface Index {
  attributes?: {
    audience_segment?: string[];
    theme_variant?: string;
    minimize_blade_spacing_bottom?: boolean;
    minimize_blade_spacing?: boolean;
    experiment_id?: string;
    experiment_type?: string;
    headline: string;
    headline_font_color?: string;
    anchor_id?: string;
    card_type: string;
    card_source?: string;
    blade_background_type?: Background;
    esi_blade_flag?: string;
    esi_blade_variant?: string;
    esi_blade_rule?: string;
  };
  name: "sf/index";
  id: string;
  media_items?: (CardResource | CardWebinar)[];
  media_items_count?: number;
  filters?: IndexBladeFilters;
  [k: string]: unknown;
}
export interface FilterGroup {
  type?: string;
  name?: string;
  label?: string;
  options?: FilterGroupOption[];
  [k: string]: unknown;
}
export interface FilterGroupOption {
  id?: string;
  label?: string;
  [k: string]: unknown;
}
export interface Logogrid {
  attributes?: {
    audience_segment?: string[];
    theme_variant?: string;
    minimize_blade_spacing_bottom?: boolean;
    minimize_blade_spacing?: boolean;
    experiment_id?: string;
    experiment_type?: string;
    card_type: string;
    card_link_visibility?: ToggleControl;
    grid_type?: string;
    logo_ticker_title?: string;
    eyebrow?: Eyebrow;
    headline: string;
    headline_font_color?: string;
    description?: string;
    anchor_id?: string;
    link_1?: Cta;
    link_2?: Cta;
    blade_background_type?: Background;
    esi_blade_flag?: string;
    esi_blade_variant?: string;
    esi_blade_rule?: string;
  };
  name: "sf/logogrid";
  id: string;
  media_items?: (CardLogomark | CardProduct)[];
  [k: string]: unknown;
}
export interface Marquee {
  attributes?: {
    blade_layout: string;
    two_column_spacing?: string;
    audience_segment?: string[];
    image_overflow?: boolean;
    theme_variant?: string;
    minimize_blade_spacing_bottom?: boolean;
    minimize_blade_spacing?: boolean;
    experiment_id?: string;
    experiment_type?: string;
    eyebrow?: Eyebrow;
    use_eyebrow_as_headline?: boolean;
    headline: string;
    headline_font_color?: string;
    headline_size?: string;
    description?: string;
    anchor_id?: string;
    link_1?: Cta;
    link_2?: Cta;
    media?: Media;
    blade_background_type?: Background;
    esi_blade_flag?: string;
    esi_blade_variant?: string;
    esi_blade_rule?: string;
  };
  name: "sf/marquee";
  id: string;
  [k: string]: unknown;
}
// ...
export interface PageBackground {
  kind: "page_standard_background";
  page_background_color?: string;
  page_background_image?: Image;
  page_background_image_position?: string;
  page_background_shape?: string;
  page_background_gradient_direction?: string;
}
export interface Footer {
  footerType?: string;
  [k: string]: unknown;
}